# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryMediaKind": {
        "enum": [
          "MOVIE",
          "EPISODE"
        ],
        "title": "HistoryMediaKind",
        "type": "string"
      },
      "HistoryScanActionRequest": {
        "properties": {
          "action": {
            "enum": [
              "start",
              "pause",
              "resume",
              "scan"
            ],
            "title": "Action",
            "type": "string"
          },
          "limit": {
            "default": 100,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          }
        },
        "required": [
          "action"
        ],
        "title": "HistoryScanActionRequest",
        "type": "object"
      },
      "HistoryScanBatchResponse": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "processed_count": {
            "title": "Processed Count",
            "type": "integer"
          },
          "scan": {
            "$ref": "#/components/schemas/HistoryScanResponse"
          }
        },
        "required": [
          "scan",
          "processed_count",
          "has_more"
        ],
        "title": "HistoryScanBatchResponse",
        "type": "object"
      },
      "HistoryScanCreateRequest": {
        "properties": {
          "exclude_patterns": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Exclude Patterns",
            "type": "array"
          },
          "extensions": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "minItems": 1,
            "title": "Extensions",
            "type": "array"
          },
          "media_kind": {
            "$ref": "#/components/schemas/HistoryMediaKind"
          },
          "root_path": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Root Path",
            "type": "string"
          }
        },
        "required": [
          "root_path",
          "media_kind",
          "extensions"
        ],
        "title": "HistoryScanCreateRequest",
        "type": "object"
      },
      "HistoryScanListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryScanResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "HistoryScanListResponse",
        "type": "object"
      },
      "HistoryScanResponse": {
        "properties": {
          "changed_count": {
            "title": "Changed Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "discovered_count": {
            "title": "Discovered Count",
            "type": "integer"
          },
          "exclude_patterns": {
            "items": {
              "type": "string"
            },
            "title": "Exclude Patterns",
            "type": "array"
          },
          "extensions": {
            "items": {
              "type": "string"
            },
            "title": "Extensions",
            "type": "array"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Completed At"
          },
          "last_started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Started At"
          },
          "media_kind": {
            "$ref": "#/components/schemas/HistoryMediaKind"
          },
          "new_count": {
            "title": "New Count",
            "type": "integer"
          },
          "root_path": {
            "title": "Root Path",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/HistoryScanStatus"
          },
          "unchanged_count": {
            "title": "Unchanged Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "root_path",
          "media_kind",
          "extensions",
          "exclude_patterns",
          "status",
          "generation",
          "cursor",
          "discovered_count",
          "new_count",
          "changed_count",
          "unchanged_count",
          "version",
          "last_started_at",
          "last_completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "HistoryScanResponse",
        "type": "object"
      },
      "HistoryScanStatus": {
        "enum": [
          "READY",
          "SCANNING",
          "PAUSED",
          "DONE"
        ],
        "title": "HistoryScanStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PackBreaker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/history-scans": {
      "get": {
        "operationId": "list_history_scans_api_v1_history_scans_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "packbreaker_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Packbreaker Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryScanListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List History Scans",
        "tags": [
          "history-scans"
        ]
      },
      "post": {
        "operationId": "create_history_scan_api_v1_history_scans_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "packbreaker_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Packbreaker Session"
            }
          },
          {
            "in": "cookie",
            "name": "packbreaker_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Packbreaker Csrf"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryScanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create History Scan",
        "tags": [
          "history-scans"
        ]
      }
    },
    "/api/v1/history-scans/{scan_id}": {
      "get": {
        "operationId": "get_history_scan_api_v1_history_scans__scan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "packbreaker_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Packbreaker Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History Scan",
        "tags": [
          "history-scans"
        ]
      }
    },
    "/api/v1/history-scans/{scan_id}/actions": {
      "post": {
        "operationId": "history_scan_action_api_v1_history_scans__scan_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "packbreaker_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Packbreaker Session"
            }
          },
          {
            "in": "cookie",
            "name": "packbreaker_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Packbreaker Csrf"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryScanActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HistoryScanResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HistoryScanBatchResponse"
                    }
                  ],
                  "title": "Response History Scan Action Api V1 History Scans  Scan Id  Actions Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History Scan Action",
        "tags": [
          "history-scans"
        ]
      }
    },
...
  }
}